+++ src/compas_timber/connections/__init__.py
@@ -8,10 +8,13 @@ from .l_miter import MiterPlaneSpec
 from .l_french_ridge_lap import LFrenchRidgeLapJoint
 from .lap_joint import LapJoint
 from .joint_candidate import JointCandidate
-from .joint_candidate import PlateJointCandidate
+from .topology_data import TopologyData
+from .topology_data import BeamTopologyData
+from .topology_data import PlateTopologyData
 from .solver import ConnectionSolver
 from .solver import PlateConnectionSolver
 from .solver import JointTopology
+from .solver import SolverResult
 from .solver import find_neighboring_elements
 from .candidate_dispatch import get_connection_candidate
 from .t_butt import TButtJoint
@@ -62,11 +65,14 @@ __all__ = [
     "TLapJoint",
     "LLapJoint",
     "JointCandidate",
-    "PlateJointCandidate",
+    "TopologyData",
+    "BeamTopologyData",
+    "PlateTopologyData",
     "LFrenchRidgeLapJoint",
     "JointTopology",
     "ConnectionSolver",
     "PlateConnectionSolver",
+    "SolverResult",
     "find_neighboring_elements",
     "get_connection_candidate",
     "TDovetailJoint",